+++ src/main/java/Logica/Fhabitacion.java
@@ -110,7 +110,7 @@ public class Fhabitacion {
            pst.setString(2, dts.getPiso());
            pst.setString(3, dts.getDescripcion());
            pst.setString(4, dts.getCaracteristicas());
-           pst.setDouble(5, dts.getPrecio());
+           pst.setInt(5, dts.getPrecio());
            pst.setString(6, dts.getEstado());
            pst.setString(7, dts.getTipohabitacion());
            
@@ -288,8 +288,10 @@ public String obtenerEstadoHabitacion(int idhabitacion) {
 
         return estado;
     }
-    
-  
+
+
+
+        
 
  public Map<Integer, String> obtenerEstadosTodasHabitaciones() {
         Map<Integer, String> estados = new HashMap<>();
